refactor: moved helper cli tests to separate file
