+++ hpcflow/tests/unit/test_helper.py
@@ -10,8 +10,6 @@ import sys
 import subprocess
 import psutil
 
-from click.testing import CliRunner
-
 from hpcflow.sdk.helper import helper
 from tempfile import gettempdir
 from hpcflow.api import hpcflow, load_config
@@ -423,7 +421,7 @@ def test_run_helper_detects_parameter_changes(app):
 # TODO: test_helper_cli (or should this be a separate test file?)
 
 
-# TODO: The tests below are actually functional tests... move them to another folder?
+# TODO: The test below is actually a functional test... move to another folder?
 def test_modify_helper(app):
     helper.clear_helper(app)
     start_t = datetime.now()
@@ -471,66 +469,3 @@ def test_modify_helper(app):
     assert update_count == 3
     assert mod_count == 2
     helper.clear_helper(app)
-
-
-def test_modify_helper_cli(app):
-    start_t = datetime.now()
-    time.sleep(0.2)
-    r = CliRunner()
-
-    so = cli(r, args="helper clear")
-    so = cli(
-        r, args="helper start --timeout 60 --timeout-check-interval 1 --watch-interval 3"
-    )
-    assert "Helper started successfully." in so
-    so = cli(
-        args="helper modify --timeout 60 --timeout-check-interval 1 --watch-interval 3"
-    )
-    assert so == "Helper parameters already met."
-
-    so = cli(
-        r, args="helper modify --timeout 60 --timeout-check-interval 2 --watch-interval 1"
-    )
-    assert so == ""
-    time.sleep(3.5)
-    so = cli(
-        r, args="helper modify --timeout 60 --timeout-check-interval 2 --watch-interval 1"
-    )
-    assert so == "Helper parameters already met."
-
-    so = cli(
-        r, args="helper modify --timeout 10 --timeout-check-interval 2 --watch-interval 1"
-    )
-    assert so == ""
-    time.sleep(3)
-    so = cli(
-        r, args="helper modify --timeout 10 --timeout-check-interval 2 --watch-interval 1"
-    )
-    assert so == "Helper parameters already met."
-
-    time.sleep(5)
-    so = cli(r, args="helper pid")
-    assert so == "Helper not running!"
-
-    read_logs = helper.read_helper_log(app, start_t)
-    mod_count = 0
-    update_count = 0
-    timeout = 0
-    for line in read_logs:
-        if " - INFO - " in line:
-            (t, m) = line.split(" - INFO - ")
-            if "Modifying" in m:
-                mod_count = mod_count + 1
-            elif "Updated" in m:
-                update_count = update_count + 1
-            elif "Helper exiting due to timeout" in m:
-                timeout = timeout + 1
-    assert timeout == 1
-    assert update_count == 3
-    assert mod_count == 2
-    so = cli(r, args="helper clear")
-
-
-def cli(r=CliRunner(), args=""):
-    so = r.invoke(hpcflow.CLI, args)
-    return so.output.strip()
